\name{BridgeR-package}
\alias{BridgeR-package}
\alias{BridgeR}
\docType{package}
\title{
An R library for comprehensive BRIC-seq data analysis tool
}
\description{
Comprehensive BRIC-seq data analysis tool
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BridgeR}
\packageIndices{BridgeR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Naoto Imamachi

Maintainer: Naoto Imamachi
}
\references{
https://github.com/Naoto-Imamachi/BRIC-seq_data_analysis/tree/master/BridgeR
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ BRIC-seq }
\seealso{

}
\examples{
#~~ simple examples of the most important functions ~~
}
