% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphic.R
\name{createGraphic}
\alias{createGraphic}
\title{createGraphic}
\usage{
createGraphic(handNo, handN, handE, handS, handW, dealer, vuln, points)
}
\arguments{
\item{handNo}{The id of the hand}

\item{handN}{The North hand generated by bridgeHand}

\item{handE}{The East hand generated by bridgeHand}

\item{handS}{The South hand generated by bridgeHand}

\item{handW}{The West hand generated by bridgeHand}

\item{dealer}{The hand to become South, the designated dealer}

\item{vuln}{The hand's vulnerability}

\item{points}{The hand's points}
}
\value{
ggplot graphic object
}
\description{
Create the graphic of the hand
}
