% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{zzz}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Legacy dummy}

\item{pkgname}{Legacy dummy}
}
\value{
No return value, called to set global variables and specify import packages
}
\description{
Runs on loading bridger
}
