% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_1major}
\alias{find_1major}
\title{find_1major}
\usage{
find_1major(HC_low = 15, HC_high = 19, cardLen_min = 4, canape = FALSE)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{HC_high}{The maximum number of high-card points, otherwise 2-level bid is possible}

\item{cardLen_min}{The minimum number of cards in the major}

\item{canape}{Whether a 4 card major will be opened before a 5 card minor}
}
\value{
id and seat of compliant hand
}
\description{
Return a bridge hand that will open 1 of a major

 Assumes that a 5 card minor will be bid before 4 card major, except if "canape" set to TRUE, then a 6 card minor
 will be opened before a 4 card major

 Assumes a weak 1NT, so HC_low is the first point outside the range of 1NT.
}
