% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCompositeHands.R
\name{find_1major_jacoby2NT}
\alias{find_1major_jacoby2NT}
\title{find_1major_jacoby2NT}
\usage{
find_1major_jacoby2NT(HC_low = 13, cardLen_low = 4)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}
}
\value{
id and seat of a compliant hand
}
\description{
Find hands where South opens one of a major, and North will bid 2NT, to show 4 card support and points for game
}
