% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_2preempt}
\alias{find_2preempt}
\title{find_2preempt}
\usage{
find_2preempt(HC_low = 5, HC_high = 10, cardLen_low = 6, cardLen_high = 7)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{HC_high}{The maximum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}

\item{cardLen_high}{The maximum length of a suit}
}
\value{
id and seat of compliant hand
}
\description{
Find hands that are likely to preempt at the 2 level in a major
}
