% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_3preempt}
\alias{find_3preempt}
\title{find_3preempt}
\usage{
find_3preempt(HC_low = 5, HC_high = 10, cardLen_low = 7, cardLen_high = 8)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{HC_high}{The maximum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}

\item{cardLen_high}{The maximum length of a suit}
}
\value{
FALSE if not compliant, or id and seat of compliant hand
}
\description{
Find hands that are likely to preempt at the 3 level
}
