% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_4441}
\alias{find_4441}
\title{find_4441}
\usage{
find_4441(HC_low = 12, HC_high = 35, cardLen_low = 5, cardLen_high = 13)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{HC_high}{The maximum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}

\item{cardLen_high}{The maximum length of a suit}
}
\value{
id and seat of compliant hand
}
\description{
Find hands that comply with a 4441 shape and opening point count
}
