% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_any}
\alias{find_any}
\title{find_any}
\usage{
find_any(HC_low = 12)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}
}
\value{
FALSE if not compliant, or id and seat of compliant hand
}
\description{
Return any bridge hand - Default
}
