% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_opener}
\alias{find_opener}
\title{find_opener}
\usage{
find_opener(HC_low = 12)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}
}
\value{
id and seat of compliant hand
}
\description{
Return a bridge hand that is likely to open
}
