% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_strong}
\alias{find_strong}
\title{find_strong}
\usage{
find_strong(HC_low = 19, HC_high = 35, cardLen_low = 1, cardLen_high = 5)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{HC_high}{The maximum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}

\item{cardLen_high}{The maximum length of a suit}
}
\value{
id and seat of compliant hand
}
\description{
Find hands that are strong enough to open strong
}
