% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHandFinders.R
\name{find_strongNT}
\alias{find_strongNT}
\title{find_strongNT}
\usage{
find_strongNT(HC_low = 15, HC_high = 17, cardLen_low = 2, cardLen_high = 5)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{HC_high}{The maximum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}

\item{cardLen_high}{The maximum length of a suit}
}
\value{
id and seat of compliant hand
}
\description{
Find hands that comply with a weak no trump opening
}
