% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCompositeHands.R
\name{find_weak1NT_LHObid}
\alias{find_weak1NT_LHObid}
\title{find_weak1NT_LHObid}
\usage{
find_weak1NT_LHObid(HC_low = 7, cardLen_low = 6)
}
\arguments{
\item{HC_low}{The minimum number of high-card points}

\item{cardLen_low}{The minimum length of a suit}
}
\value{
id and seat of a compliant hand
}
\description{
Find hands where South will open a weak 1NT and West will likely bid
}
