% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
No return value, called to allow piping
}
\description{
## Implements the magrittr Pipe function
}
\keyword{internal}
