% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyPlays.R
\name{safety_plays}
\alias{safety_plays}
\title{safety_plays}
\usage{
safety_plays(cards_n = "AKTxx", cards_s = "xxxx", smallCards = 9)
}
\arguments{
\item{cards_n}{The cards North holds}

\item{cards_s}{The cards South holds}

\item{smallCards}{Cards to be considered small and equivalent, default = 9}
}
\value{
data.frame of all possible layouts of the four hands
}
\description{
Show all combinations of all remaining cards to assess safety plays.
  Next stage would be to automatically assessment to achieve required number of tricks.
}
\examples{
safety_plays()

safety_plays(smallCards = 10)

}
