% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeHand.R
\name{bridgeHand}
\alias{bridgeHand}
\title{bridgeHand}
\usage{
bridgeHand(handNumber = "auto", seat = FALSE, createGraphic = TRUE, ...)
}
\arguments{
\item{handNumber}{An integer for generating a hand, or "auto" to use a random number generator}

\item{seat}{If not false, makes the specified seat South and dealer, so all bidding starts with South and the specified hand type}

\item{createGraphic}{Whether the graphic should be created}

\item{...}{Other parameters passed to shape the hands}
}
\value{
List: Hand ID, Dealer, Hand graphic, Hand points, Hand shape
}
\description{
Generate a bridge hand
}
\note{
To change the scoring of hands pass high card values (HCValues) and hand shape values (shapeValues) in the arguments.

HCValues is a string of five digits specifying the value of the Ace, King, Queen, Jack and 10.  The default is , 3, 2, 1, 0.
shapeValues is a string of eight digits specifying the value of a suit with no cards/"Void", 1-card/"Singleton", ... 7-cards.  The default is c(3, 2, 1, 0, 0, 1, 2, 3)
}
\examples{
# Produce a bridge hand
hand <- bridgeHand()

# Produce a bridge hand 500 ensuring South as dealer
hand500 <- bridgeHand(handNumber = "500", seat = "S") # Seat can be any compass-point

# Produce a bridge hand, with South as dealer, but without generating the graphic
handIDonly <- bridgeHand(seat = "W", createGraphic = FALSE)
}
