\name{BridgeR-package}
\alias{BridgeR-package}
\alias{BridgeR}
\docType{package}
\title{
\packageTitle{BridgeR}
}
\description{
\packageDescription{BridgeR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BridgeR}
\packageIndices{BridgeR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BridgeR}

Maintainer: \packageMaintainer{BridgeR}
}
\references{
https://github.com/Naoto-Imamachi/BRIC-seq_data_analysis/tree/master/BridgeR
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{

}
\examples{
#~~ simple examples of the most important functions ~~
}
