% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printHands.R
\name{printHands}
\alias{printHands}
\title{printHands}
\usage{
printHands(handType = "any", num = 12, ...)
}
\arguments{
\item{handType}{The type of hand required, default is 'any'.  Alternatives include, '4441', 'strong', ...}

\item{num}{The number of hands wanted - Make divisible by 6}

\item{...}{Other variables that may be passed when selecting compliant hands}
}
\description{
Produce a graphic of hands on a single page, useful as a hand-out.  Each page can hold upto 6 hands.
}
\examples{
# Produce 2 pages of 6 hands of bridge hands
printHands()

# Produce a page of 6 hands likely to open with a 3-level preempt
printHands(num = 6, handType = "preempt3")
}
