% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printHands.R
\name{printHands}
\alias{printHands}
\title{printHands}
\usage{
printHands(
  ids = FALSE,
  seats = FALSE,
  handType = "any",
  num = 12,
  output = "F",
  saveOutputDir = "temp",
  ...
)
}
\arguments{
\item{ids}{The ids of hands to be generated}

\item{seats}{The seats of the hands in ids, i.e. the seat which gives the requested conditions}

\item{handType}{The type of hand required, default is 'any'.  Alternatives include, '4441', 'strong', ...}

\item{num}{The number of hands wanted}

\item{output}{Character code of required seats, "N", "E", "S", "W" and "F" for the full hand NB "ALL" equivalent to "FNEWS"}

\item{saveOutputDir}{Default "temp" to save to temporary directory, or specify a directory, e.g. "c:/temp/bridger"}

\item{...}{Other variables that may be passed when selecting compliant hands}
}
\description{
Produce a graphic of hands as a hand-out.  Each page can hold up to 6 hands.

  Pages of full hands, or only one set of hands can be selected.  "ALL" is equivalent to FULL and separate pages of
  each of the four positions.  "N", "E", "S" or "W" will produce only show the points and hands for those positions.

  The output will be a PDF for the selected hand sets.  The default is to save them to a default directory, but a directory can be specified
  in saveOutput in which the PDF will be saved.
}
\examples{
\dontrun{
# Produce a hand showing all seats and save them to c:/temp/bridger
printHands(handType = "any", num = 1, output = "all", saveOutput = "c:/temp/bridger")

# Produce a page of 6 South hands likely to open with a 3-level preempt
printHands(handType = "preempt3", num = 6, output = "S")

# Produce specified hands
printHands(ids = c(500, 501, 502), seats = c("E", "W", "S", output = "F"))
}

}
