% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printHands.R
\name{printHands}
\alias{printHands}
\title{printHands}
\usage{
printHands(
  ids = FALSE,
  seats = FALSE,
  handType = "any",
  num = 12,
  output = "F",
  saveOutputDir = "c:/temp/bridger/",
  ...
)
}
\arguments{
\item{ids}{The ids of hands to be generated}

\item{seats}{The seats of the hands in ids, i.e. the seat which gives the requested conditions}

\item{handType}{The type of hand required, default is 'any'.  Alternatives include, '4441', 'strong', ...}

\item{num}{The number of hands wanted}

\item{output}{Character code of required outputs, "N", "E", "S", "W" for the respective hands, "F" for the full hand NB "ALL" equivalent to "FNEWS"}

\item{saveOutputDir}{The location for saving the output pages, Set to NULL to save to a temporary directory}

\item{...}{Other variables that may be passed when selecting compliant hands}
}
\description{
Produce a graphic of hands as a hand-out.  Each page can hold up to 6 hands.

  Pages of full hands, or only one set of hands can be selected.  "ALL" is equivalent to FULL and separate pages of
  each of the four positions.  "N", "E", "S" or "W" will produce only show the points and hands for those positions.

  The output will be a PDF for the selected hand sets.  These will be saved to the 'c:/temp/bridger/' directory, if saveOutput
  is set to TRUE, or to temporary, notified files if set to false.
}
\examples{
# Produce 1 hand showing only one seat
printHands(handType = "any", num = 1, output = "ALL")

\dontrun{
# Produce a page of 6 South hands likely to open with a 3-level preempt
    printHands(handType = "preempt3", num = 6, output = "S")

# Produce specified hands
    printHands(ids = c(500, 501, 502), seats = c("E", "W", "S"))
}

}
