% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitSplit.R
\name{suitSplit}
\alias{suitSplit}
\title{suitSplit}
\usage{
suitSplit(missingCards = 5, cards_W = 13, cards_E = 13)
}
\arguments{
\item{missingCards}{The number of cards held by the two hands}

\item{cards_W}{Cards in West hands, e.g. If bidding indicates another 5 card suit, for example by overcalling, then reduce by cards_W by 5}

\item{cards_E}{Cards in East hands}
}
\value{
Tibble of probabilities
}
\description{
The probabilities of how a number of cards  will split between two hands, given a number of unknown cards in each hand.  Unknown hands
  are assumed to be West and East.
}
\note{
If the unknown cards in both hands are the same, then symmetrical probabilities will be returned.
}
\examples{
suitSplit(missingCards = 6, cards_W = 13, cards_E = 8)

}
